refactored: navReducer (for types and rarities) moved to file filterReducer and exported out
