+++ src/components/Nav.js
@@ -1,14 +1,15 @@
 import React from "react";
 import { Link } from "react-router-dom";
 import { useSelector, useDispatch } from 'react-redux';
-import { getRarities, getTypes } from "../store";
+import { getRarities, getTypes, selectRarities, selectTypes } from "../features/filterReducer";
+import { selectProducts } from "../features/productsReducer";
 
 const Nav = () => {
     const dispatch = useDispatch();
-    const products = useSelector(state => state.products);
+    const products = useSelector(selectProducts);
     // gets an array of types and rarities from the state after first render
-    let types = useSelector(state => state.nav.types);
-    let rarities = useSelector(state => state.nav.rarities);
+    let types = useSelector(selectTypes);
+    let rarities = useSelector(selectRarities);
 
     React.useEffect(() => {
         dispatch(getTypes());
